TESTS: arm32: Fix compile errors for FreeRTOS on VExpress-A9 Guest
The FreeRTOS for VExpress-A9 compilation was broken due to recent
re-factoring in basic firmware. This patch fixes the compile errors.

+++ tests/arm32/common/basic/arch_types.h
@@ -31,7 +31,7 @@ typedef unsigned short u16;
 typedef unsigned int u32;
 typedef unsigned int size_t;
 typedef unsigned int bool;
-typedef unsigned int ulong;
+typedef unsigned long ulong;
 
 /** Boolean macros */
 #define TRUE		1
